Add App Icon
This icon was generated with Bakery. I am also working on an icon to fit the Scribbler vibe. This commit also removed some title bar coloring, because that messed with the custom font. We can color it... but then it doesn't fill the whole way. Need to talk to a UI expert.